Removing one unnecessory import/move createQuery,createQueriesList,cleanUp to utils
