+++ auto_nag/scripts/rm_query_creator.py
@@ -4,10 +4,10 @@ import requests
 import re
 import datetime
 import subprocess
-import os
 import json
 from argparse import ArgumentParser
-from auto_nag.bugzilla.utils import get_config_path, get_project_root_path
+from auto_nag.bugzilla.utils import (get_config_path, get_project_root_path,
+                                     createQueriesList, cleanUp)
 
 
 def getTemplateValue(url):
@@ -32,40 +32,6 @@ urls = [
     (5, ["Notify release managers when bugs are marked fixed in nightly but still affected for aurora, beta or release", "fixed_without_uplifts", fixed_without_uplifts_url, 0])
 ]
 
-
-def createQuery(queries_dir, title, short_title, url, show_summary):
-    file_name = queries_dir + str(datetime.date.today()) + '_' + short_title
-    if not os.path.exists(queries_dir):
-        os.makedirs(queries_dir)
-    qf = open(file_name, 'w')
-    qf.write("query_name = \'" + title + "\'\n")
-    qf.write("query_url = \'" + url + "\'\n")
-    qf.write("show_summary = \'" + str(show_summary) + "\'\n")
-    return file_name
-
-
-def createQueriesList(queries_dir, weekday, print_all):
-    queries = []
-    for url in urls:
-        if weekday >= 0 and weekday < 5 and url[0] == 5:
-            queries.append(createQuery(queries_dir, title=url[1][0], short_title=url[1][1], url=url[1][2], show_summary=url[1][3]))
-        if weekday == 0 and url[0] == 0:
-            queries.append(createQuery(queries_dir, title=url[1][0], short_title=url[1][1], url=url[1][2], show_summary=url[1][3]))
-        if weekday == 3 and url[0] == 3:
-            queries.append(createQuery(queries_dir, title=url[1][0], short_title=url[1][1], url=url[1][2], show_summary=url[1][3]))
-    print queries
-    return queries
-
-
-def cleanUp(queries_dir):
-    try:
-        for file in os.listdir(queries_dir):
-            if file.startswith(str(datetime.date.today())):
-                os.remove(os.path.join(queries_dir, file))
-        return True
-    except Exception as error:
-        print "Error: ", str(error)
-        return False
 
 if __name__ == '__main__':
     # basic setups
